refactor(js/test): break up test file so it is possible to execute only a few at a time
