+++ crypto-ffi/bindings/js/test/initialization.test.ts
@@ -1,0 +1,20 @@
+import {
+    ccInit,
+    setup,
+    teardown,
+} from "./utils.js";
+import { afterEach, beforeEach, describe } from "mocha";
+
+beforeEach(async () => {
+    await setup();
+});
+
+afterEach(async () => {
+    await teardown();
+});
+
+describe("initialization", () => {
+    it("should succeed", async () => {
+        await ccInit("foo");
+    });
+});
